requirements.txt
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
grpcio==1.64.1
idna==3.7
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
msgpack==1.0.8
numpy==2.0.0
packaging==24.1
pandas==2.2.2
protobuf==5.27.2
pyarrow==16.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
ray==2.32.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
